fix(deps): remove criterion from library crate graphs (#150)
- Move vec_znx Criterion bench helpers from poulpy-cpu-ref to poulpy-bench
- Drop unused criterion dependency from poulpy-cpu-ref, poulpy-hal, poulpy-core, poulpy-schemes
- Reword poulpy-hal crate docs to point at poulpy-bench for tests/benches
- Refresh Cargo.lock
