+++ poulpy-bench/src/bench_suite/hal/vec_znx/mod.rs
@@ -1,0 +1,22 @@
+//! Criterion benchmark harnesses for coefficient-domain [`VecZnx`] operations.
+//!
+//! Each submodule mirrors the corresponding `poulpy-cpu-ref` `vec_znx` implementation
+//! file; the helpers are invoked from `poulpy-bench` harness binaries.
+
+mod add;
+mod automorphism;
+mod mul_xp_minus_one;
+mod negate;
+mod normalize;
+mod rotate;
+mod shift;
+mod sub;
+
+pub use add::{bench_vec_znx_add_assign, bench_vec_znx_add_into};
+pub use automorphism::{bench_vec_znx_automorphism, bench_vec_znx_automorphism_inplace};
+pub use mul_xp_minus_one::{bench_vec_znx_mul_xp_minus_one, bench_vec_znx_mul_xp_minus_one_inplace};
+pub use negate::{bench_vec_znx_negate, bench_vec_znx_negate_inplace};
+pub use normalize::{bench_vec_znx_normalize, bench_vec_znx_normalize_inplace};
+pub use rotate::{bench_vec_znx_rotate, bench_vec_znx_rotate_inplace};
+pub use shift::{bench_vec_znx_lsh, bench_vec_znx_lsh_inplace, bench_vec_znx_rsh, bench_vec_znx_rsh_inplace};
+pub use sub::{bench_vec_znx_sub, bench_vec_znx_sub_inplace, bench_vec_znx_sub_negate_inplace};
