+++ poulpy-cpu-ref/Cargo.toml
@@ -13,7 +13,6 @@ documentation = "https://docs.rs/poulpy"
 poulpy-core = {workspace = true}
 poulpy-hal = {workspace = true}
 dashu-float = {workspace = true}
-criterion = {workspace = true}
 bytemuck = {workspace = true}
 itertools = {workspace = true}
 rand = {workspace = true}
