+++ poulpy-cpu-ref/src/reference/vec_znx/sub.rs
@@ -1,12 +1,6 @@
-use std::hint::black_box;
-
-use criterion::{BenchmarkId, Criterion};
-
 use crate::{
-    api::{ModuleNew, VecZnxSub, VecZnxSubInplace, VecZnxSubNegateInplace},
-    layouts::{Backend, FillUniform, Module, VecZnx, VecZnxToMut, VecZnxToRef, ZnxInfos, ZnxView, ZnxViewMut},
+    layouts::{VecZnx, VecZnxToMut, VecZnxToRef, ZnxInfos, ZnxView, ZnxViewMut},
     reference::znx::{ZnxCopy, ZnxNegate, ZnxNegateInplace, ZnxSub, ZnxSubInplace, ZnxSubNegateInplace, ZnxZero},
-    source::Source,
 };
 
 pub fn vec_znx_sub<R, A, B, ZNXARI>(res: &mut R, res_col: usize, a: &A, a_col: usize, b: &B, b_col: usize)
@@ -113,140 +107,4 @@ where
     for j in sum_size..res_size {
         ZNXARI::znx_negate_inplace(res.at_mut(res_col, j));
     }
-}
-
-pub fn bench_vec_znx_sub<B>(c: &mut Criterion, label: &str)
-where
-    B: Backend,
-    Module<B>: VecZnxSub + ModuleNew<B>,
-{
-    let group_name: String = format!("vec_znx_sub::{label}");
-
-    let mut group = c.benchmark_group(group_name);
-
-    fn runner<B: Backend>(params: [usize; 3]) -> impl FnMut()
-    where
-        Module<B>: VecZnxSub + ModuleNew<B>,
-    {
-        let n: usize = 1 << params[0];
-        let cols: usize = params[1];
-        let size: usize = params[2];
-
-        let module: Module<B> = Module::<B>::new(n as u64);
-
-        let mut source: Source = Source::new([0u8; 32]);
-
-        let mut a: VecZnx<Vec<u8>> = VecZnx::alloc(n, cols, size);
-        let mut b: VecZnx<Vec<u8>> = VecZnx::alloc(n, cols, size);
-        let mut c: VecZnx<Vec<u8>> = VecZnx::alloc(n, cols, size);
-
-        // Fill a with random i64
-        a.fill_uniform(50, &mut source);
-        b.fill_uniform(50, &mut source);
-
-        move || {
-            for i in 0..cols {
-                module.vec_znx_sub(&mut c, i, &a, i, &b, i);
-            }
-            black_box(());
-        }
-    }
-
-    for params in [[10, 2, 2], [11, 2, 4], [12, 2, 8], [13, 2, 16], [14, 2, 32]] {
-        let id: BenchmarkId = BenchmarkId::from_parameter(format!("{}x({}x{})", 1 << params[0], params[1], params[2],));
-        let mut runner = runner::<B>(params);
-        group.bench_with_input(id, &(), |b, _| b.iter(&mut runner));
-    }
-
-    group.finish();
-}
-
-pub fn bench_vec_znx_sub_inplace<B>(c: &mut Criterion, label: &str)
-where
-    B: Backend,
-    Module<B>: VecZnxSubInplace + ModuleNew<B>,
-{
-    let group_name: String = format!("vec_znx_sub_inplace::{label}");
-
-    let mut group = c.benchmark_group(group_name);
-
-    fn runner<B: Backend>(params: [usize; 3]) -> impl FnMut()
-    where
-        Module<B>: VecZnxSubInplace + ModuleNew<B>,
-    {
-        let n: usize = 1 << params[0];
-        let cols: usize = params[1];
-        let size: usize = params[2];
-
-        let module: Module<B> = Module::<B>::new(n as u64);
-
-        let mut source: Source = Source::new([0u8; 32]);
-
-        let mut a: VecZnx<Vec<u8>> = VecZnx::alloc(n, cols, size);
-        let mut b: VecZnx<Vec<u8>> = VecZnx::alloc(n, cols, size);
-
-        // Fill a with random i64
-        a.fill_uniform(50, &mut source);
-        b.fill_uniform(50, &mut source);
-
-        move || {
-            for i in 0..cols {
-                module.vec_znx_sub_inplace(&mut b, i, &a, i);
-            }
-            black_box(());
-        }
-    }
-
-    for params in [[10, 2, 2], [11, 2, 4], [12, 2, 8], [13, 2, 16], [14, 2, 32]] {
-        let id: BenchmarkId = BenchmarkId::from_parameter(format!("{}x({}x{})", 1 << params[0], params[1], params[2]));
-        let mut runner = runner::<B>(params);
-        group.bench_with_input(id, &(), |b, _| b.iter(&mut runner));
-    }
-
-    group.finish();
-}
-
-pub fn bench_vec_znx_sub_negate_inplace<B>(c: &mut Criterion, label: &str)
-where
-    B: Backend,
-    Module<B>: VecZnxSubNegateInplace + ModuleNew<B>,
-{
-    let group_name: String = format!("vec_znx_sub_negate_inplace::{label}");
-
-    let mut group = c.benchmark_group(group_name);
-
-    fn runner<B: Backend>(params: [usize; 3]) -> impl FnMut()
-    where
-        Module<B>: VecZnxSubNegateInplace + ModuleNew<B>,
-    {
-        let n: usize = 1 << params[0];
-        let cols: usize = params[1];
-        let size: usize = params[2];
-
-        let module: Module<B> = Module::<B>::new(n as u64);
-
-        let mut source: Source = Source::new([0u8; 32]);
-
-        let mut a: VecZnx<Vec<u8>> = VecZnx::alloc(n, cols, size);
-        let mut b: VecZnx<Vec<u8>> = VecZnx::alloc(n, cols, size);
-
-        // Fill a with random i64
-        a.fill_uniform(50, &mut source);
-        b.fill_uniform(50, &mut source);
-
-        move || {
-            for i in 0..cols {
-                module.vec_znx_sub_negate_inplace(&mut b, i, &a, i);
-            }
-            black_box(());
-        }
-    }
-
-    for params in [[10, 2, 2], [11, 2, 4], [12, 2, 8], [13, 2, 16], [14, 2, 32]] {
-        let id: BenchmarkId = BenchmarkId::from_parameter(format!("{}x({}x{})", 1 << params[0], params[1], params[2]));
-        let mut runner = runner::<B>(params);
-        group.bench_with_input(id, &(), |b, _| b.iter(&mut runner));
-    }
-
-    group.finish();
 }
